requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.16
botocore==1.31.16
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==3.2.20
django-angular @ git+https://github.com/jrief/django-angular.git@ae9d77786f79d13cde3696ec35012feff02654fe
django-anymail==10.0
-e git+https://github.com/maykonmenezes/django-bcp.git@6ce7dd35cfc9c09961b363bddd1e89a35d6196ae#egg=django_bcp
django-bootstrap4==23.2
django-currentuser==0.5.3
django-filter==23.2
django-form-renderers==0.3.2
django-heroku==0.3.1
django-import-export==3.2.0
django-localflavor==4.0
django-logentry-admin==1.1.0
django-parsley==0.7
django-storages==1.13.2
django-widget-tweaks==1.4.12
docutils==0.20.1
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
jmespath==0.10.0
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Parsley==1.3
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-social-auth==0.2.12
python-stdnum==1.18
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uuid==1.30
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
